+++ arrow-core-data/src/test/kotlin/arrow/core/EvalTest.kt
@@ -1,6 +1,7 @@
 package arrow.core
 
 import arrow.Kind
+import arrow.core.computations.eval
 import arrow.core.extensions.eq
 import arrow.core.extensions.eval.applicative.applicative
 import arrow.core.extensions.eval.bimonad.bimonad
@@ -38,7 +39,7 @@ class EvalTest : UnitSpec() {
 
     testLaws(
       BimonadLaws.laws(Eval.bimonad(), Eval.monad(), Eval.comonad(), Eval.functor(), Eval.applicative(), Eval.monad(), GENK, EQK),
-      FxLaws.laws<ForEval, Int>(G, G, EQK.liftEq(Int.eq()), evalFx::eager, evalFx::invoke)
+      FxLaws.laws<ForEval, Int>(G, G, EQK.liftEq(Int.eq()), eval::eager, eval::invoke)
     )
 
     "should map wrapped value" {
